fix: Align billing tests with catalog rates
The Business plan change lowered VM catalog rates and updated trial
copy. Tests still expected the old hosted-credit strings, raw machine
ids, and a 90-second VM cost of at least one cent.

Record five minutes of large-x64 usage so VM cents stay positive.
Update onboarding and usage tests to the new copy and labels. Split
the billing page so ESLint stays inside the line budget.
