+++ pkg/grpc/actions/factories/export_factory_work_order_run_usage_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/superplanehq/superplane/pkg/database"
 	"github.com/superplanehq/superplane/pkg/models"
 	pb "github.com/superplanehq/superplane/pkg/protos/factories"
+	"github.com/superplanehq/superplane/pkg/usage/pricebook"
 	"github.com/superplanehq/superplane/test/support"
 )
 
@@ -59,14 +60,22 @@ func Test__ExportFactoryWorkOrderRunUsage(t *testing.T) {
 		InputTokens:     1_000_000,
 		TotalTokens:     1_000_000,
 	}))
+	const (
+		computeMachineType = "e1-large-amd64"
+		computeSeconds     = int64(300)
+	)
+	computeCostCents := pricebook.MicrosToCents(
+		pricebook.EstimateComputeMicros(computeMachineType, computeMachineType, computeSeconds),
+	)
+	require.Greater(t, computeCostCents, int64(0))
 	require.NoError(t, models.RecordComputeUsage(db, models.ComputeUsageEventInput{
 		OrganizationID:  r.Organization.ID,
 		CanvasRunID:     *execution.RunID,
 		NodeExecutionID: uuid.New(),
 		NodeID:          "runner",
-		MachineType:     "e1-large-amd64",
-		FleetID:         "e1-large-amd64",
-		DurationSeconds: 90,
+		MachineType:     computeMachineType,
+		FleetID:         computeMachineType,
+		DurationSeconds: computeSeconds,
 		IdempotencyKey:  "runner:compute:export-usage:" + uuid.New().String(),
 	}))
 
@@ -99,16 +108,13 @@ func Test__ExportFactoryWorkOrderRunUsage(t *testing.T) {
 	assert.Equal(t, factory.WorkOrderKey(order.Number), row[2])
 	assert.Equal(t, "anthropic/claude-sonnet-4-6 (your keys)", row[3])
 	assert.Equal(t, "1000000", row[4])
-	assert.Equal(t, "e1-large-amd64", row[6])
-	assert.Equal(t, "90", row[7])
+	assert.Equal(t, computeMachineType, row[6])
+	assert.Equal(t, strconv.FormatInt(computeSeconds, 10), row[7])
 
 	tokenPriceCents, err := strconv.ParseFloat(row[5], 64)
 	require.NoError(t, err)
 	assert.Greater(t, tokenPriceCents, 0.0)
-
-	vmPriceCents, err := strconv.ParseFloat(row[8], 64)
-	require.NoError(t, err)
-	assert.Greater(t, vmPriceCents, 0.0)
+	assert.Equal(t, formatUsageCSVCents(computeCostCents), row[8])
 }
 
 func Test__ExportFactoryWorkOrderRunUsage__NoRows(t *testing.T) {
