+++ web_src/src/pages/factories/pages/organizationSettings/OrganizationSettingsBillingPage.tsx
@@ -1,4 +1,4 @@
-import { Check, ChevronRight, ExternalLink, Receipt } from "lucide-react";
+import { ChevronRight } from "lucide-react";
 import { useState, type ReactNode } from "react";
 import { useParams } from "react-router";
 
@@ -8,7 +8,6 @@ import type {
   OrganizationsOrganizationCreditGrant,
 } from "@/api-client";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import { useReportPageReady } from "@/hooks/useReportPageReady";
 import { usePageTitle } from "@/hooks/usePageTitle";
 import { getApiErrorMessage } from "@/lib/errors";
@@ -24,9 +23,10 @@ import {
   type BillingCreditBucketView,
 } from "../../lib/billingCreditBuckets";
 import { hostedCreditBillingBalanceCopy } from "../../lib/hostedCreditEmpty";
-import { creditGrantDetails, creditGrantSourceLabel, formatCreditGrantAmount } from "../../lib/hostedCreditGrants";
 import { formatUsdCents, parseWorkOrderMetric } from "../../lib/workOrderUsage";
 import { FactorySettingsCard, FactorySettingsPageFrame } from "../settings/FactorySettingsCard";
+import { BillingCreditHistoryCard } from "./BillingCreditHistoryCard";
+import { BillingInvoicesCard } from "./BillingInvoicesCard";
 import { BillingPlansSection } from "./BillingPlansSection";
 import { useOrganizationBillingPageModel } from "./useOrganizationBillingPageModel";
 
@@ -34,9 +34,6 @@ const BUY_MORE_PACK_CENTS = [5_000, 10_000, 50_000] as const;
 const HOSTED_CREDIT_EXPLANATION =
   "Hosted credit pays SuperPlane-hosted machines and managed models for this organization.";
 const HOSTED_CREDIT_REMAINING_CAPTION = "Remaining hosted credit";
-const INVOICES_DESCRIPTION = "Recent paid invoices for this organization.";
-const INVOICES_EMPTY_TITLE = "No invoices yet";
-const INVOICES_EMPTY_BODY = "Paid invoices appear here after checkout.";
 
 export function OrganizationSettingsBillingPage() {
   const { organizationId = "" } = useParams<{ organizationId: string }>();
@@ -182,9 +179,9 @@ function BillingPageBody({
         onAddCredit={onAddCredit}
       />
       {showInvoices ? (
-        <PolarInvoicesCard invoices={invoices} portalPending={portalPending} onManageInvoices={onManageInvoices} />
+        <BillingInvoicesCard invoices={invoices} portalPending={portalPending} onManageInvoices={onManageInvoices} />
       ) : null}
-      <CreditHistoryCard grants={grants} />
+      <BillingCreditHistoryCard grants={grants} />
     </>
   );
 }
@@ -442,157 +439,6 @@ function HostedCreditTopUpBanner({
   );
 }
 
-function PolarInvoicesCard({
-  invoices,
-  portalPending,
-  onManageInvoices,
-}: {
-  invoices: OrganizationsHostedCreditInvoice[];
-  portalPending: boolean;
-  onManageInvoices: () => void | Promise<void>;
-}) {
-  return (
-    <FactorySettingsCard
-      title="Invoices"
-      data-testid="billing-polar-invoices"
-      action={
-        <Button
-          type="button"
-          size="sm"
-          variant="ghost"
-          disabled={portalPending}
-          onClick={() => void onManageInvoices()}
-        >
-          {portalPending ? "Opening invoices..." : "Manage invoices"}
-          <ExternalLink className="size-3.5" aria-hidden />
-        </Button>
-      }
-    >
-      <p className="text-[12px] text-muted-foreground">{INVOICES_DESCRIPTION}</p>
-      {invoices.length === 0 ? <InvoiceEmptyState /> : <InvoiceList invoices={invoices} />}
-    </FactorySettingsCard>
-  );
-}
-
-function InvoiceEmptyState() {
-  return (
-    <div className="mt-4 flex flex-col items-center rounded-lg border border-dashed border-border px-6 py-8 text-center">
-      <Receipt className="size-5 text-muted-foreground" aria-hidden />
-      <p className="mt-2 text-sm font-medium text-foreground">{INVOICES_EMPTY_TITLE}</p>
-      <p className="mt-1 text-xs text-muted-foreground">{INVOICES_EMPTY_BODY}</p>
-    </div>
-  );
-}
-
-function InvoiceList({ invoices }: { invoices: OrganizationsHostedCreditInvoice[] }) {
-  return (
-    <ul className="mt-3 divide-y divide-border">
-      {invoices.map((invoice) => (
-        <InvoiceRow key={invoice.id} invoice={invoice} />
-      ))}
-    </ul>
-  );
-}
-
-function InvoiceRow({ invoice }: { invoice: OrganizationsHostedCreditInvoice }) {
-  const productName = invoice.productName || "Hosted credit";
-  const amount = formatUsdCents(parseWorkOrderMetric(invoice.amountCents));
-  const dateLabel = formatInvoiceDate(invoice.createdAt);
-
-  return (
-    <li className="flex items-center gap-3 py-3 first:pt-1 last:pb-0">
-      <div className="flex size-8 shrink-0 items-center justify-center rounded-md bg-muted">
-        <Receipt className="size-3.5 text-muted-foreground" aria-hidden />
-      </div>
-      <div className="min-w-0 flex-1">
-        <p className="truncate text-[13px] font-medium tracking-[-0.01em]">{productName}</p>
-        {invoice.createdAt ? (
-          <time className="mt-0.5 block text-xs text-muted-foreground" dateTime={invoice.createdAt}>
-            {dateLabel}
-          </time>
-        ) : (
-          <p className="mt-0.5 text-xs text-muted-foreground">{dateLabel}</p>
-        )}
-      </div>
-      <p className="shrink-0 text-[13px] font-medium tracking-[-0.01em] tabular-nums">{amount}</p>
-      <InvoiceStatusBadge status={invoice.status} />
-    </li>
-  );
-}
-
-function InvoiceStatusBadge({ status }: { status: string | undefined }) {
-  const label = invoiceStatusLabel(status);
-  if (status === "paid") {
-    return (
-      <Badge
-        variant="outline"
-        className="border-emerald-500/30 bg-emerald-500/10 text-emerald-700 dark:text-emerald-400"
-      >
-        <Check aria-hidden />
-        {label}
-      </Badge>
-    );
-  }
-
-  return (
-    <Badge variant="outline" className="text-muted-foreground">
-      {label}
-    </Badge>
-  );
-}
-
-function CreditHistoryCard({ grants }: { grants: OrganizationsOrganizationCreditGrant[] }) {
-  return (
-    <FactorySettingsCard title="Credit history" data-testid="billing-credit-history">
-      {grants.length === 0 ? (
-        <p className="text-sm text-muted-foreground">No credit grants yet.</p>
-      ) : (
-        <CreditGrantTable grants={grants} />
-      )}
-    </FactorySettingsCard>
-  );
-}
-
-function CreditGrantTable({ grants }: { grants: OrganizationsOrganizationCreditGrant[] }) {
-  return (
-    <table className="mt-1 w-full text-left text-[13px]">
-      <thead>
-        <tr className="border-b border-border text-muted-foreground">
-          <th className="py-2 pr-3 font-medium">Date</th>
-          <th className="py-2 pr-3 font-medium">Source</th>
-          <th className="py-2 pr-3 font-medium">Amount</th>
-          <th className="py-2 font-medium">Details</th>
-        </tr>
-      </thead>
-      <tbody>
-        {grants.map((grant) => (
-          <tr key={grant.id ?? `${grant.kind}-${grant.createdAt}`} className="border-b border-border last:border-0">
-            <td className="py-2 pr-3 whitespace-nowrap">{formatGrantDate(grant.createdAt)}</td>
-            <td className="py-2 pr-3">
-              <Badge variant="secondary">{creditGrantSourceLabel(grant.kind)}</Badge>
-            </td>
-            <td className="py-2 pr-3 whitespace-nowrap">
-              {formatCreditGrantAmount(parseWorkOrderMetric(grant.amountCents))}
-            </td>
-            <td className="py-2 text-muted-foreground">{creditGrantDetails(grant) || "—"}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-}
-
-function formatInvoiceDate(value: string | undefined) {
-  if (!value) {
-    return "—";
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return value;
-  }
-  return parsed.toLocaleDateString(undefined, { month: "short", day: "numeric", year: "numeric" });
-}
-
 function findPackForCents(packs: OrganizationsHostedCreditProduct[], cents: number) {
   return packs.find((product) => Boolean(product.id) && parseWorkOrderMetric(product.amountCents) === cents);
 }
@@ -610,28 +456,4 @@ function creditRefreshClassName(status: HostedCreditRefreshStatus) {
     return "text-emerald-700 dark:text-emerald-400";
   }
   return "text-muted-foreground";
-}
-
-function formatGrantDate(value: string | undefined) {
-  if (!value) {
-    return "—";
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return value;
-  }
-  return parsed.toLocaleDateString();
-}
-
-function invoiceStatusLabel(status: string | undefined) {
-  switch (status) {
-    case "paid":
-      return "Paid";
-    case "refunded":
-      return "Refunded";
-    case "partially_refunded":
-      return "Partially refunded";
-    default:
-      return status || "Unknown";
-  }
 }
